requirements.txt
selenium
beautifulsoup4
jdatetime
psycopg2-binary
khayyam
ipython
django-extensions
jupyter
aiohttp
webdriver-manager
appdirs
mteb
notebook
transformers
torch
sentencepiece
openai
datasets
pandas
